[156] Move test extensions to their own package
